Add MessageID to a couple of fields that we want to get responses from
As it turns out, we're actually emulating a sort of RPC interface
between the broker and the coordinator. I think we can make it work, but
there are some simplifications, namely that we can only really have one
outstanding response for each message. Right now, sending a
BrokerPublishMessage to the coordinator results in both a ForwardRequest
and a SubscriptionDiff, which is hard to account for in the broker-side
code. In this case, a possibility is that all producers forward automatically
to the coordinator, which throws them away if it doesn't need them. Or,
we can combine the two messages into one?