Adding in BrokerIDs and fixing Broker addresses
Making sure that we are exporting the right information to the
coordinator, including the UUID for the broker and the actual address it
uses. Right now, I'm using the "client" side of the broker's address and
sending that around because its easier. We can still handle all of the
acknowledgements, etc for BrokerPublish vs regular Publish using the
"client" side interface, though right now I'm just using regular
Publish